requirements.txt
torch==2.0.1
transformers==4.29.2
sense2vec==2.0.2
strsim==0.0.3
six==1.16.0
networkx==3.1
numpy==1.22.4
scipy==1.10.1
scikit-learn==1.2.2
unidecode==1.3
future==0.18.3
joblib==1.2.0
pytz==2022.7.1
python-dateutil==2.8.2
flashtext==2.7
pandas==1.5.3
sentencepiece==0.1.99
nltk>=3.8.1
wget
spacy>=3.7.2
sacremoses
python-docx
openpyxl